users page refactoring, paginator
