+++ src/components/Content/Users/User.jsx
@@ -6,47 +6,40 @@ import { NavLink } from "react-router-dom";
 
 
 
-const User = ({ user, followingInProgress, unfollow, follow}) => {
+const User = ({ user, followingInProgress, unfollow, follow }) => {
 
     return (
-        <div >
-                       
-                <div  className={css.usersElement}>
+        <div className={css.usersElement}>
+            <div >
+                <NavLink to={"/profile/" + user.id} >
+                    <img alt='avatar' src={
+                        user.photos.small != null
+                            ? user.photos.small
+                            : userPhoto} className={css.imageUsers} />
+                </NavLink>
+            </div>
+            <div className={css.usersData}>
+                <div>Name: {user.name}</div>
+                <div>ID:{user.id}</div>
+                <div>{"user.locaction.country"}</div>
+                <div>Status:{user.status}</div>
+                <div>{"user.locaction.city"}</div>
+            </div >
+            <div className={css.button}>
+                {user.followed
+                    ? <button disabled={followingInProgress
+                        .some(id => id === user.id)} className={css.unfollow}
+                        onClick={() => { unfollow(user.id) }}>
+                        Unfollow {user.name}</button>
 
-                        <div >
-                            <NavLink to={"/profile/" + user.id} >
-                                <img alt='avatar' src={
-                                    user.photos.small != null
-                                        ? user.photos.small
-                                        : userPhoto} className={css.imageUsers} />
-                            </NavLink>
-                        </div>
-                        <div className={css.usersData}>
-                            <div>Name: {user.name}</div>
-                            <div>ID:{user.id}</div>
-                            <div>{"user.locaction.country"}</div>
-                            <div>Status:{user.status}</div>
-                            <div>{"user.locaction.city"}</div>
-                        </div >
-                        <div className={css.button}>
-                            {user.followed
-                            
-                                ? <button disabled={followingInProgress
-                                    .some(id => id === user.id)} className={css.unfollow}
-                                    onClick={() => {unfollow(user.id) }}>
-                                    Unfollow {user.id}</button>
-
-                                : <button disabled={followingInProgress
-                                    .some(id => id === user.id)} className={css.follow}
-                                    onClick={() => {follow(user.id) }}>
-                                    Follow me! {user.id}</button>
-                            }
-                        </div>
-                    </div>
-                    
-                
-           
+                    : <button disabled={followingInProgress
+                        .some(id => id === user.id)} className={css.follow}
+                        onClick={() => { follow(user.id) }}>
+                        Follow {user.name}</button>
+                }
+            </div>
         </div>
+
 
     )
 
@@ -56,12 +49,3 @@ const User = ({ user, followingInProgress, unfollow, follow}) => {
 
 export default User;
 
-
-{/* {users.map(u => <User user={u}
-                followingInProgress={props.followingInProgress}
-                unfollow={props.unfollow}
-                follow={props.follow}
-                key={u.id}
-
-            />
-            )} */}
