+++ src/components/common/Paginator/Paginator.module.css
@@ -1,0 +1,5 @@
+.sectedPage {
+    font-weight: bold;
+    font-size: large;
+    color: var(--redC);
+}
